refactor: split the probe-capture cluster out of world-json-types
The file held ~110 lines of probe-capture types beside ClassifierStatus, the one genuine JSON column in it — and the capture payload is not a JSON column at all, it is a gzipped BLOB. DropReason moved with the cluster: it is capture and retrieval vocabulary rather than column shape, and the capture payload is what persists it.

The name now means what it says, and world-json-types.ts is left with no imports at all — which takes @/lib/retrieval off the transitive graph of stories.table.ts, the file that pulls ClassifierStatus. Consumers are unaffected: everything reaches these through the @/lib/db barrel, so only its re-export paths moved.
