+++ lib/db/index.ts
@@ -41,13 +41,9 @@ export { translationWriteSchema } from './translations/translations-schema'
 export type { TranslationWrite } from './translations/translations-schema'
 export { branchEraFlipWriteSchema } from './stories/era-flip-schema'
 export type { BranchEraFlipWrite } from './stories/era-flip-schema'
-export { CAPTURE_VERSION } from './world-json-types'
-export type {
-  ClassifierLifecycleState,
-  ClassifierStatus,
-  DropReason,
-  ProbeCapturePayload,
-} from './world-json-types'
+export { CAPTURE_VERSION } from './probe-capture-types'
+export type { DropReason, ProbeCapturePayload } from './probe-capture-types'
+export type { ClassifierLifecycleState, ClassifierStatus } from './world-json-types'
 export type {
   StoryDefinition,
   StoryMode,
